Tweaks and fixes from 1.8.9, fixes mega item registration on servers

+++ Pokecube Core/src/main/java/pokecube/core/entity/pokemobs/helper/EntityMovesPokemob.java
@@ -298,7 +298,7 @@ public abstract class EntityMovesPokemob extends EntitySexedPokemob
     @Override
     public int getExplosionState()
     {
-        return dataWatcher.get(BOOMSTATEDW);
+        return (int) dataWatcher.get(BOOMSTATEDW);
     }
 
     public int getLastAttackTick()
